+++ laravel/documentation/changes.md
@@ -54,6 +54,7 @@
 - Allow multiple views to be registered for a single composer.
 - Added `Request::set_env` method.
 - `Schema::drop` now accepts `$connection` as second parameter.
+- Added `Input::merge` method.
 
 <a name="upgrade-3.2"></a>
 ## Upgrading From 3.1